Refactor code for clarity and maintainability
This update includes changes that improve readability and maintainability of code. Command code is broken down into smaller, reusable utility functions. Constant values and regular expressions are defined at the top of files. Loops and conditionals have been simplified. This overall leads to code that is easier to understand and modify.