Add QR code sending and update payment properties
This commit introduces the functionality of sending a QR code to the interaction channelId. A dependency on the primary module is also added at the beginning of the file. Additionally, some minor updates were made to the properties of the payment transaction, including parsing the product price to a number type and adding a notification URL.